Switch to using wsgi.py as wsgi/django.wsgi has been deprecated
The following is included in django.wsgi:

```
Use of this 'djano.wsgi' file has been deprecated since the Rocky
release in favor of 'wsgi.py' in the 'openstack_dashboard' module. This
file is a legacy naming from before Django 1.4 and an importable
'wsgi.py' is now the default. This file will be removed in the T release
cycle.
```

Change-Id: Ic188e6ec87cb59d9311787fa36defd193116c2b0
Signed-off-by: Lance Albertson <lance@osuosl.org>